% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gangsta_queryGangstas.R
\name{subsetGangstas}
\alias{subsetGangstas}
\alias{getGangstas}
\alias{getGangstaAttribute}
\alias{gangstasExist}
\title{Query a list of \code{gangsta} objects}
\usage{
subsetGangstas(gangstaObjects, attributeName, attributeValue,
  asIndex = F)

getGangstas(gangstaObjects, gangstaNames)

getGangstaAttribute(gangstaObjects, attribName)

gangstasExist(gangstaObjects, gangstaNames, checkClass = "")
}
\arguments{
\item{gangstaObjects}{List of \code{gangsta} objects to be queried.}

\item{attributeName}{Name of the attribute used for the query.}

\item{attributeValue}{All \code{gangsta} objects where the attribute
(specified in \code{attributeName}) is equal to \code{attributeValue} will
be returned.}

\item{asIndex}{Causes return value to be a logical vector that can be applied
to \code{gangstaObjects}.}

\item{gangstaNames}{Character vector containing the names (value stored in
the \code{name} attribute) of the \code{gangsta} objects to be returned.}

\item{attribName}{Name of an attribute associated with 1 or more
\code{gangsta} classes.}

\item{checkClass}{When set to a value other than "", checks to be sure all of
the objects referenced in \code{gangstaNames} are of the class
\code{checkClass}. Throws an error if not.}
}
\value{
\code{gangstasExist} returns TRUE if all names in \code{gangstaNames}
  have an associated \code{gangsta} object in \code{gangstaObjects}.  If not,
  execution is halted and an error reported.  \code{getGangstaAttribute}
  returns a vector of attribute values from the objects in
  \code{gangstaObjects}.  All other functions return a \code{list} of
  \code{gangsta} objects.
}
\description{
Lists of \code{gangsta} objects created using \code{compoundFactory} and
\code{processFactory} can be queried conveniently using these methods.
}
\details{
\code{subsetGangstas} Queries \code{gangsta} objects by attribute.  Returns a
subset of \code{gangstaObjects} where the attribute specified by
\code{attributeName} is equal to the value specified by
\code{attributeValue}.

\code{gangstasExist} returns \code{TRUE} if all names in \code{gangstaNames}
have an associated \code{gangsta} object in \code{gangstaObjects}.  If not,
execution is halted and an error reported.

\code{getGangstas} is a simplified wrapper for \code{subsetGangstas} and
queries \code{gangstaObjects} on the "name" attribute.  It also throws an
error if no object of a provided name exists in \code{gangstaObjects} or if
more than one object with a provided name exists in \code{gangstaObjects}.

\code{getGangstaAttribute} returns a vector of attribute values from the
objects in \code{gangstaObjects}. Usually this is most useful if
\code{gangstaObjects} is set equal to results returned from
\code{getGangstas} or \code{subsetGangstas}.
}
