% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gangsta_queryGangstas.R
\name{subsetGangstas}
\alias{subsetGangstas}
\alias{getGangstas}
\alias{getGangstaAttribute}
\alias{gangstasExist}
\title{Query a list of GANGSTA Objects}
\usage{
subsetGangstas(gangstaObjects, attributeName, attributeValue)

getGangstas(gangstaObjects, gangstaNames)

getGangstaAttribute(gangstaObjects, attribName)

gangstasExist(gangstaObjects, gangstaNames, checkClass = "")
}
\arguments{
\item{gangstaObjects}{List of GANGSTA objects to be queried.}

\item{attributeName}{Name of the attributed used for the query.}

\item{attributeValue}{All GANGSTA objects where the attribute (specified in
\code{attribureName}) is equal to \code{attributeValue} will be returned.}

\item{gangstaNames}{Character vector containing the names (value stored in
the \code{name} attribute) of the GANGSTA objects to be returned.}

\item{attribName}{Name of an attribute associated with 1 or more GANGSTA
classes.}

\item{checkClass}{When set to a value other than "", checks to be sure all of
the objects referenced in \code{gangstaNames} are of the class
\code{checkClass}. Throws an error if not.}
}
\value{
\code{gangstasExist} returns TRUE if all names in \code{gangstaNames}
  have an associated GANGSTA object in \code{gangstaObjects}.  If not,
  execution is halted and an error reported.  \code{getGangstaAttribute}
  returns a vector of attribute values from the objects in
  \code{gangstaObjects}.  All other functions return a \code{list} of GANGSTA
  objects.
}
\description{
Lists of GANGSTA objects created using \code{compoundFactory} and
\code{processFactory} can be queried conveniently using these
methods.
}
\details{
\code{subsetGangstas} Queries gangstaObjects by attribute.  Returns a subset
of \code{gangstaObjects} where the attribute specified by
\code{attributeName} is equal to the value specified by
\code{attributeValue}.

\code{gangstasExist} returns TRUE if all names in \code{gangstaNames} have an
associated GANGSTA object in \code{gangstaObjects}.  If not, execution is
halted and an error reported.

\code{getGangstas} is a simplified wrapper for \code{subsetGangstas} and
queries \code{gangstaObjects} on the "name" attributure.  It also throws an
error if no object of a provided name exists in \code{gangstaObjects} or if
more than one object with a provided name exists in \code{gangstaObjects}.

\code{getGangstaAttribute} returns a vector of attribute values from the
objects in \code{gangstaObjects}. Ususally this is most useful if
\code{gangstaObjects} is set equal to results returned from
\code{getGangstas} or \code{subsetGangstas}.
}
